requirements.txt
Django==1.11.17
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1